check service exist em service executed

+++ src/modules/services/useCases/deleteServiceById/DeleteServiceByIdUseCase.ts
@@ -1,6 +1,7 @@
 import { ensureId } from "@ensures/ensureId";
 import { ICreateServiceDTO } from "@modules/services/dtos/ICreateServiceDTO";
 import { IServicesRepository } from "@modules/services/repositories/IServicesRepository";
+import { IServiceExecutedRepository } from "@modules/servicesExcuteds/repositories/IServiceExecutedRepository";
 import { inject, injectable } from "tsyringe";
 
 import { AppError } from "@shared/errors/AppError";
@@ -9,7 +10,9 @@ import { AppError } from "@shared/errors/AppError";
 export class DeleteServiceByIdUseCase {
     constructor(
         @inject("ServicesRepository")
-        private servicesRepository: IServicesRepository
+        private servicesRepository: IServicesRepository,
+        @inject("ServiceExecutedRepository")
+        private serviceExecutedRepository: IServiceExecutedRepository
     ) {}
 
     async execute({ id }: ICreateServiceDTO): Promise<void> {
@@ -17,10 +20,22 @@ export class DeleteServiceByIdUseCase {
             throw new AppError("Service not found", 404);
         }
 
-        const checkServiceExists = this.servicesRepository.findById(id);
+        const checkServiceExists = await this.servicesRepository.findById(id);
 
         if (!checkServiceExists) {
             throw new AppError("Service not found", 404);
+        }
+
+        const checkExistServiceExcuted =
+            await this.serviceExecutedRepository.listByServiceId(
+                checkServiceExists.id
+            );
+
+        if (checkExistServiceExcuted.length > 0) {
+            throw new AppError(
+                "Cannot delete service when there is a related executed service",
+                404
+            );
         }
 
         await this.servicesRepository.deleteById(id);
